+++ src/beta/openai/chat.rs
@@ -3,6 +3,9 @@
 *   Copyright © 2026 NatML Inc. All Rights Reserved.
 */
 
+use std::sync::Arc;
+
+use crate::client::Client;
 use crate::services::{PredictionService, PredictorService};
 
 use super::ChatCompletionService;
@@ -17,11 +20,12 @@ pub struct ChatService {
 impl ChatService {
 
     pub fn new(
+        client: Arc<dyn Client>,
         predictors: PredictorService,
         predictions: PredictionService
     ) -> Self {
         Self {
-            completions: ChatCompletionService::new(predictors, predictions),
+            completions: ChatCompletionService::new(client, predictors, predictions),
         }
     }
 }