+++ src/beta/openai/mod.rs
@@ -15,6 +15,9 @@ pub use embeddings::*;
 pub use images::*;
 pub use schema::*;
 
+use std::sync::Arc;
+
+use crate::client::Client;
 use crate::services::{PredictionService, PredictorService};
 
 /// Experimental OpenAI client.
@@ -31,10 +34,11 @@ pub struct OpenAIClient {
 impl OpenAIClient {
 
     pub fn new(
+        client: Arc<dyn Client>,
         predictors: PredictorService,
         predictions: PredictionService
     ) -> Self {
-        let chat = ChatService::new(predictors.clone(), predictions.clone());
+        let chat = ChatService::new(client, predictors.clone(), predictions.clone());
         let embeddings = EmbeddingService::new(predictors.clone(), predictions.clone());
         let images = ImageService::new(predictors, predictions);
         Self { chat, embeddings, images }